--- a/package.json
+++ b/package.json
@@ -7,7 +7,7 @@
     "al": "index.js"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "al -a x @ echo testing && al x && al -l && al -r x"
   },
   "keywords": [
     "alias",