--- a/src/opts.js
+++ b/src/opts.js
@@ -3,8 +3,8 @@ const { alias, options } = require('./config')
 const flags = [
   ['--version, -v', alias.version],
   ['--help, -h', alias.help],
-  ['--add, -a', alias.addAlias],
-  ['--remove, -r', alias.removeAlias],
+  ['--add, -a', alias.add],
+  ['--remove, -r', alias.remove],
   ['--list, -l', alias.list],
   ['--separator, -s', options.setSeparator]
 ]