the following script:

```js
const getConfig = require('./src/config.js').getConfig
console.log(getConfig())
```

logs the `defaultConfig` value, not the config in my home directory. although:

```js
const { getConfig, checkConfig } = require('./src/config.js')

checkConfig(() => {
  console.log(getConfig())
})
```

logs the correct user config

this is because of the callback-based architecture - although, since alia is a command line tool and doesn't need to handle multiple asynchronous events at any one time (like a server would), the callbacks make the code a bit more awkward to work with (imo) while not saving any run-time.

this pr just removes callbacks and handles file read/writes synchronously

### speed tests

running the following test script as `time ./test.sh`

```bash
al -a t @ echo test
al t
al -r t
```

over 5 test runs before and after pr changes, gave the following results

| test run       | master /ms | this pr /ms |
|---------|--------|---------|
| 1       | 365    | 376     |
| 2       | 437    | 319     |
| 3       | 349    | 384     |
| 4       | 334    | 379     |
| 5       | 400    | 377     |
| Average | 377    | 367     |

as you can see, no time is lost (infact, pr change test runs were slightly faster but not significantly)