@edkotkas 2018-10-16
The time difference is insignificant and the callback will ensure anything ran in the application will have the correct config file loaded.