@jmsv 2018-10-16
yeah, the point I was making _was_ that the time diff is insignificant

using synchronous read/writes ensures config loads without the need for callbacks