@edkotkas 2018-10-16
What do you think of this? Using existsync (seems not to be deprecated yet)
https://stackoverflow.com/questions/4482686/check-synchronously-if-file-directory-exists-in-node-js