@jmsv 2018-10-16
check update commit

using `existsync` along with `fs.lstatSync(configPath).isFile()` to check it's not a directory