@jmsv 2018-10-17
reverted previous commit and updated `loadConfig`